fix: rename package from pockertracker to pokertracker (typo fix)
