+++ android/app/src/main/AndroidManifest.xml
@@ -25,7 +25,7 @@
         <category android:name="android.intent.category.DEFAULT"/>
         <category android:name="android.intent.category.BROWSABLE"/>
         <data android:scheme="myapp"/>
-        <data android:scheme="com.yonatantseitlin.pockertracker"/>
+        <data android:scheme="com.yonatantseitlin.pokertracker"/>
       </intent-filter>
     </activity>
   </application>
